changed R naming scheme (ugh)
It's that stupid thing about . in names.  I was using dots because
they're used in function names of packages I use.  But it sporadically
causes trouble with roxygen2 thinking it means S3 method (e.g.,
scale.root.age).  There is a workaround (add to @export and @usage
lines), but that's ugly and asking for trouble in the future.

Anyway, the idea now is for all the phycombR code to use snake case.
The only downside to underscores is inconsistency with names in other
packages we rely on.  Well, there's also the issue with emacs/vim
shortcuts for ->, but that seems smaller-scale and re-map-able.