first Methods, Elements, Results
Elements are assembled from component Trees and Traits.  File names are
standardized in this process.

Methods are provided as scripts.  There can be multiple files of code,
but there needs to be one function of the correct format to run the
method and report the result.

Then there is a script to run a given Method on a given Element,
producing a Result file.  This script should maybe eventually be in
phycombR, but it isn't yet.

A contributor could upload Elements, Methods, and/or Results.  After the
upload, in the database, Methods will be stored as code files, Elements
may be stored as-is or maybe just the assembly information, and Results
will probably be imported and not saved as csv files.